+++ CHANGELOG.md
@@ -4,13 +4,13 @@
 
 ## v0.0.4
 
-(TODO: date)
+2025-01-23
 
 - `*` Wallet: Dash, DigiByte and Zcash address derivation and message signing
 - `*` Wallet: Solana and Polkadot/Kusama/Substrate address derivation
 - `*` Wallet: Auto-lock after inactivity
 - `+` Wallet: improved password hashing (breaking change!)
-- `#` Core: tested and fixed prebuilt board configs (TODO:)
+- `#` Core: tested and fixed prebuilt board configs
 
 ## v0.0.3